server/requirements.txt
openenv-core
fastapi
uvicorn
numpy
openai
ortools
python-dotenv